Moved common tests into their own package.
Also includes a some new find by id and name common tests.
